+++ db/migrate/20180507204653_rename_slug_to_survey_slug_for_designs.rb
@@ -1,0 +1,5 @@
+class RenameSlugToSurveySlugForDesigns < ActiveRecord::Migration[5.2]
+  def change
+    rename_column :designs, :slug, :survey_slug
+  end
+end