+++ db/migrate/20180507211109_add_slug_again_to_designs.rb
@@ -1,0 +1,6 @@
+class AddSlugAgainToDesigns < ActiveRecord::Migration[5.2]
+  def change
+    add_column :designs, :slug, :string
+    add_index :designs, [:project_id, :slug], unique: true
+  end
+end