+++ test/controllers/survey_controller_test.rb
@@ -15,7 +15,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should get new survey with slug" do
-    get new_survey_path(slug: @public_design.slug)
+    get new_survey_path(slug: @public_design.survey_slug)
     assert_not_nil assigns(:project)
     assert_not_nil assigns(:design)
     assert_equal true, assigns(:design).publicly_available
@@ -24,7 +24,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
 
   test "should not get new private survey" do
     assert_equal false, @private_design.publicly_available
-    get new_survey_path(slug: @private_design.slug)
+    get new_survey_path(slug: @private_design.survey_slug)
     assert_nil assigns(:project)
     assert_nil assigns(:design)
     assert_redirected_to about_survey_path
@@ -34,7 +34,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_difference("SheetTransaction.count") do
       assert_difference("Subject.count") do
         assert_difference("Sheet.count") do
-          post create_survey_path(slug: designs(:admin_public_design).slug)
+          post create_survey_path(slug: designs(:admin_public_design).survey_slug)
         end
       end
     end
@@ -45,14 +45,14 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:sheet)
     assert_not_nil assigns(:sheet).authentication_token
     assert_not_nil assigns(:sheet).last_edited_at
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug, a: assigns(:sheet).authentication_token)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug, a: assigns(:sheet).authentication_token)
   end
 
   test "should not submit public survey without required fields" do
     assert_difference("SheetTransaction.count", 0) do
       assert_difference("Subject.count", 0) do
         assert_difference("Sheet.count", 0) do
-          post create_survey_path(designs(:admin_public_design_with_required_fields).slug), params: {
+          post create_survey_path(designs(:admin_public_design_with_required_fields).survey_slug), params: {
             subject_id: subjects(:external).id,
             variables: { variables(:public_autocomplete).id.to_s => "" }
           }
@@ -70,7 +70,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
   test "should submit public survey and redirect to redirect_url" do
     assert_difference("Subject.count") do
       assert_difference("Sheet.count") do
-        post create_survey_path(slug: designs(:admin_public_design_with_redirect).slug)
+        post create_survey_path(slug: designs(:admin_public_design_with_redirect).survey_slug)
       end
     end
     assert_not_nil assigns(:design)
@@ -86,7 +86,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_difference("SheetTransaction.count") do
       assert_difference("Subject.count") do
         assert_difference("Sheet.count") do
-          post create_survey_path(slug: designs(:admin_public_design).slug), params: {
+          post create_survey_path(slug: designs(:admin_public_design).survey_slug), params: {
             site_id: ""
           }
         end
@@ -95,14 +95,14 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:sheet)
     assert_not_nil assigns(:sheet).subject
     assert_equal sites(:admin_site).id, assigns(:sheet).subject.site_id
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug, a: assigns(:sheet).authentication_token)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug, a: assigns(:sheet).authentication_token)
   end
 
   test "should submit public survey with first site selected" do
     assert_difference("SheetTransaction.count") do
       assert_difference("Subject.count") do
         assert_difference("Sheet.count") do
-          post create_survey_path(slug: designs(:admin_public_design).slug), params: {
+          post create_survey_path(slug: designs(:admin_public_design).survey_slug), params: {
             site_id: sites(:admin_site).id
           }
         end
@@ -111,14 +111,14 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:sheet)
     assert_not_nil assigns(:sheet).subject
     assert_equal sites(:admin_site).id, assigns(:sheet).subject.site_id
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug, a: assigns(:sheet).authentication_token)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug, a: assigns(:sheet).authentication_token)
   end
 
   test "should submit public survey with second site selected" do
     assert_difference("SheetTransaction.count") do
       assert_difference("Subject.count") do
         assert_difference("Sheet.count") do
-          post create_survey_path(slug: designs(:admin_public_design).slug), params: {
+          post create_survey_path(slug: designs(:admin_public_design).survey_slug), params: {
             site_id: sites(:admin_site_two).id
           }
         end
@@ -127,14 +127,14 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
     assert_not_nil assigns(:sheet)
     assert_not_nil assigns(:sheet).subject
     assert_equal sites(:admin_site_two).id, assigns(:sheet).subject.site_id
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug, a: assigns(:sheet).authentication_token)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug, a: assigns(:sheet).authentication_token)
   end
 
   test "should not submit private survey" do
     assert_difference("SheetTransaction.count", 0) do
       assert_difference("Subject.count", 0) do
         assert_difference("Sheet.count", 0) do
-          post create_survey_path(slug: designs(:admin_design).slug)
+          post create_survey_path(slug: designs(:admin_design).survey_slug)
         end
       end
     end
@@ -148,7 +148,7 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
 
   test "should get edit survey using authentication_token" do
     get edit_survey_path(
-      slug: designs(:admin_public_design).slug,
+      slug: designs(:admin_public_design).survey_slug,
       sheet_authentication_token: sheets(:external).authentication_token
     )
     assert_not_nil assigns(:sheet)
@@ -158,39 +158,39 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
 
   test "should not edit sheet survey with invalid authentication_token" do
     get edit_survey_path(
-      slug: designs(:admin_public_design).slug,
+      slug: designs(:admin_public_design).survey_slug,
       sheet_authentication_token: "123"
     )
     assert_not_nil assigns(:project)
     assert_nil assigns(:sheet)
     assert_equal "This survey no longer exists.", flash[:alert]
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug)
   end
 
   test "should not edit auto-locked sheet survey" do
     get edit_survey_path(
-      slug: designs(:auto_lock).slug,
+      slug: designs(:auto_lock).survey_slug,
       sheet_authentication_token: sheets(:auto_lock).authentication_token
     )
     assert_not_nil assigns(:project)
     assert_not_nil assigns(:sheet)
     assert_equal "This survey has been locked.", flash[:alert]
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug)
   end
 
   test "should resubmit sheet survey using authentication_token" do
     patch update_survey_path(
-      slug: designs(:admin_public_design).slug,
+      slug: designs(:admin_public_design).survey_slug,
       sheet_authentication_token: sheets(:external).authentication_token
     )
     assert_not_nil assigns(:project)
     assert_not_nil assigns(:sheet)
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug, a: assigns(:sheet).authentication_token)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug, a: assigns(:sheet).authentication_token)
   end
 
   test "should not resubmit sheet survey with missing required fields" do
     patch update_survey_path(
-      slug: designs(:admin_public_design_with_required_fields).slug,
+      slug: designs(:admin_public_design_with_required_fields).survey_slug,
       sheet_authentication_token: sheets(:external_with_required_fields).authentication_token
     ), params: {
       variables: { variables(:public_autocomplete).id.to_s => "" }
@@ -205,23 +205,23 @@ class SurveyControllerTest < ActionDispatch::IntegrationTest
 
   test "should not resubmit sheet survey using invalid authentication_token" do
     patch update_survey_path(
-      slug: designs(:admin_public_design).slug,
+      slug: designs(:admin_public_design).survey_slug,
       sheet_authentication_token: "123"
     )
     assert_not_nil assigns(:project)
     assert_nil assigns(:sheet)
     assert_equal "This survey no longer exists.", flash[:alert]
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug)
   end
 
   test "should not resubmit auto-locked sheet survey" do
     patch update_survey_path(
-      slug: designs(:auto_lock).slug,
+      slug: designs(:auto_lock).survey_slug,
       sheet_authentication_token: sheets(:auto_lock).authentication_token
     )
     assert_not_nil assigns(:project)
     assert_not_nil assigns(:sheet)
     assert_equal "This survey has been locked.", flash[:alert]
-    assert_redirected_to about_survey_path(survey: assigns(:design).slug)
+    assert_redirected_to about_survey_path(survey: assigns(:design).survey_slug)
   end
 end